bot_files/pyproject.toml
[tool.poetry]
name = "feedmewell"
version = "0.1.0"
description = ""
authors = ["Artem Samoilov <samoylovartem07@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
loguru = "^0.7.0"
python-telegram-bot = "13.12"
requests = "2.28.1"
pymongo = "4.1.1"
python-dotenv = "^1.0.0"
emoji = "^1.7.0"
dnspython = "2.2.1"
pydantic = "^1.10.9"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
